+++ src/bindings/bnd_extensions.cpp
@@ -1036,12 +1036,14 @@ bool BND_File3dmLayerTable::Delete(BND_UUID id)
 BND_Layer* BND_File3dmLayerTable::FindName(std::wstring name, BND_UUID parentId)
 {
   ON_UUID id = Binding_to_ON_UUID(parentId);
-  ON_ModelComponentReference compref  = m_model->LayerFromName(id, name.c_str());
-  const ON_ModelComponent* model_component = compref.ModelComponent();
-  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(model_component));
-  if (modellayer)
-    return new BND_Layer(modellayer, &compref, m_model);
-  return nullptr;
+  ON_ModelComponentReference cr = m_model->ComponentFromName(ON_ModelComponent::Type::Layer, id, name.c_str());
+
+  if (cr.IsEmpty()){
+    return nullptr;
+  }
+
+  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(cr.ModelComponent()));
+  return new BND_Layer(modellayer, &cr, m_model);
 }
 
 BND_Layer* BND_File3dmLayerTable::IterIndex(int index)
@@ -1051,28 +1053,31 @@ BND_Layer* BND_File3dmLayerTable::IterIndex(int index)
 
 BND_Layer* BND_File3dmLayerTable::FindIndex(int index)
 {
-  ON_ModelComponentReference compref = m_model->LayerFromIndex(index);
-  const ON_ModelComponent* model_component = compref.ModelComponent();
-  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(model_component));
-  if (modellayer)
-    return new BND_Layer(modellayer, &compref, m_model);
+  ON_ModelComponentReference cr = m_model->ComponentFromIndex(ON_ModelComponent::Type::Layer, index);
 
-#if defined(ON_PYTHON_COMPILE)
-  throw py::index_error();
-#else
-  return nullptr;
-#endif
+  if (cr.IsEmpty()){
+    #if defined(ON_PYTHON_COMPILE)
+      throw py::index_error();
+    #else
+      return nullptr;
+    #endif
+  }
+
+  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(cr.ModelComponent()));
+  return new BND_Layer(modellayer, &cr, m_model);
 }
 
 BND_Layer* BND_File3dmLayerTable::FindId(BND_UUID id)
 {
   ON_UUID _id = Binding_to_ON_UUID(id);
-  ON_ModelComponentReference compref = m_model->LayerFromId(_id);
-  const ON_ModelComponent* model_component = compref.ModelComponent();
-  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(model_component));
-  if (modellayer)
-    return new BND_Layer(modellayer, &compref, m_model);
-  return nullptr;
+  ON_ModelComponentReference cr = m_model->ComponentFromId(ON_ModelComponent::Type::Layer, _id);
+
+  if (cr.IsEmpty()){
+    return nullptr;
+  }
+
+  ON_Layer* modellayer = const_cast<ON_Layer*>(ON_Layer::Cast(cr.ModelComponent()));
+  return new BND_Layer(modellayer, &cr, m_model);
 }
 
 void BND_File3dmGroupTable::Add(const BND_Group& group)